requirement.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.5.0
PyJWT==2.10.1
python-slugify==8.0.4
Pillow==9.0.0
